+++ proof-of-concept/deli-pbui-react/src/generic/clim/components/PbuiAction/index.ts
@@ -1,0 +1,2 @@
+export * from './PbuiAction';
+export * from './types';