+++ proof-of-concept/deli-pbui-react/src/generic/clim/components/PbuiAction/types.ts
@@ -1,0 +1,7 @@
+import type { ActionPresentation } from '../../types';
+
+export interface PbuiActionProps {
+  action: ActionPresentation;
+  className?: string;
+  onInvoke?: () => void;
+}