+++ proof-of-concept/deli-pbui-react/src/generic/clim/components/PbuiActionBar/types.ts
@@ -1,0 +1,7 @@
+import type { ActionPresentation } from '../../types';
+
+export interface PbuiActionBarProps {
+  actions: ActionPresentation[];
+  className?: string;
+  onInvoke?: (action: ActionPresentation) => void;
+}