+++ proof-of-concept/deli-pbui-react/src/generic/clim/components/PbuiClickableText/index.ts
@@ -1,0 +1,2 @@
+export * from './PbuiClickableText';
+export * from './types';