+++ proof-of-concept/deli-pbui-react/src/generic/clim/components/PbuiText/index.ts
@@ -1,0 +1,2 @@
+export * from './PbuiText';
+export * from './types';